1-fastapi/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
supabase
python-multipart
requests
aiofiles
openai
ffmpeg-python
python-docx

